[Hotfix] Исправлен тип данных у dt на double (#120)

+++ Libraries/PIReg/PIReg.h
@@ -13,7 +13,7 @@ private:
 	double integral_lim;
 
 public:
-	PIReg(double k_int, double k_pr, int dt, double integral_lim); //constructor
+	PIReg(double k_int, double k_pr, double dt, double integral_lim); //constructor
 	void setError(double error);
 	void calcOutput(void);
 	double getOutput(void);
